security-roles-testing

+++ journalApp/src/main/java/com/intellect/journalApp/service/JournalEntryService.java
@@ -32,7 +32,7 @@ public class JournalEntryService {
 
 
         user.getJournalEntries().add(saved);
-        userService.saveEntry(user);
+        userService.saveUser(user);
     }
 
     public void saveEntry(JournalEntry journalEntry){
@@ -47,13 +47,30 @@ public class JournalEntryService {
     public Optional<JournalEntry> findById(ObjectId id){
         return journalEntryRepository.findById(id);
     }
+    
 
     public void deleteById(ObjectId id, String userName){
-        UserEntity user = userService.findByUserName(userName);
-        user.getJournalEntries().removeIf(x->x.getId().equals(id));
-        userService.saveEntry(user);
-        journalEntryRepository.deleteById(id);
+        try {
+            UserEntity user = userService.findByUserName(userName);
+            boolean removed = user.getJournalEntries().removeIf(x -> x.getId().equals(id));
+            if(removed){
+                userService.saveUser(user);
+                journalEntryRepository.deleteById(id);
+
+            }
+
+        }
+        catch(Exception e){
+            System.out.println(e);
+            throw new RuntimeException("an errror occure: ",e);
+        }
+
+
     }
+
+//    public List<JournalEntry> findByUserName(String userName){
+//
+//    }
